[#30] Rename: postDetail 관련 컴포넌트 postDetail 폴더로 이동
