+++ src/app/board/info/[id]/page.tsx
@@ -1,4 +1,4 @@
-import PostDetail from '@/components/Board/PostDetail';
+import PostDetail from '@/components/Board/postDetail/PostDetail';
 
 export default async function InfoBoardDetailPage() {
   return (
